Fixed Makefile and warnings in cygwin/mingw
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@142 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib-util/include/pjlib-util/scanner_cis_bitwise.h
@@ -69,7 +69,7 @@ typedef struct pj_cis_t
  * @param cis       Pointer to character input specification.
  * @param c         The character.
  */
-#define PJ_CIS_SET(cis,c)   ((cis)->cis_buf[(c)] |= (1 << (cis)->cis_id))
+#define PJ_CIS_SET(cis,c)   ((cis)->cis_buf[(int)(c)] |= (1 << (cis)->cis_id))
 
 /**
  * Remove the membership of the specified character.
@@ -78,7 +78,7 @@ typedef struct pj_cis_t
  * @param cis       Pointer to character input specification.
  * @param c         The character to be removed from the membership.
  */
-#define PJ_CIS_CLR(cis,c)   ((cis)->cis_buf[c] &= ~(1 << (cis)->cis_id))
+#define PJ_CIS_CLR(cis,c)   ((cis)->cis_buf[(int)c] &= ~(1 << (cis)->cis_id))
 
 /**
  * Check the membership of the specified character.
@@ -87,7 +87,7 @@ typedef struct pj_cis_t
  * @param cis       Pointer to character input specification.
  * @param c         The character.
  */
-#define PJ_CIS_ISSET(cis,c) ((cis)->cis_buf[c] & (1 << (cis)->cis_id))
+#define PJ_CIS_ISSET(cis,c) ((cis)->cis_buf[(int)c] & (1 << (cis)->cis_id))
